+++ src/lang/nodes/Function.js
@@ -1,8 +1,13 @@
 export function compile(node, ctx) {
+    const args = node.arguments.map(arg => '$' + arg.name);
+
+    // use function(){} since Arg1 refers to arguments[1],
+    // but Function doesn't create 2nd argument implicitly to prevent function arity changes
+    ctx.put('function(');
+    ctx.put(String(args) || '$');
+
     ctx.createScope(
         () => {
-            const args = node.arguments.map(arg => '$' + arg.name);
-
             ctx.scope.arg1 = true;
             ctx.scope.$ref = args[0] || '$';
 
@@ -10,16 +15,15 @@ export function compile(node, ctx) {
                 ctx.scope.add(arg.name);
             }
 
-            ctx.put('function(');
-            ctx.put(String(args) || '$');
             ctx.put('){return ');
             ctx.node(node.body);
-            ctx.put('}');
         },
         (scopeStart, sp) => {
             return scopeStart + sp + ',';
         }
     );
+
+    ctx.put('}');
 }
 export function walk(node, ctx) {
     ctx.node(node.body);