+++ src/lang/nodes/Function.js
@@ -1,5 +1,12 @@
 export function compile(node, ctx) {
     const args = node.arguments.map(arg => '$' + arg.name);
+    const duplicateNameNode = node.arguments.find((node, index) =>
+        index > 0 && args.lastIndexOf('$' + node.name, index - 1) !== -1
+    );
+
+    if (duplicateNameNode) {
+        ctx.error('Duplicate parameter name "$' + duplicateNameNode.name + '" is not allowed', duplicateNameNode);
+    }
 
     // use function(){} since Arg1 refers to arguments[1],
     // but Function doesn't create 2nd argument implicitly to prevent function arity changes