+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Fixed a compilation error for the generated code of `function` expressions in stat mode in some cases
 - Improved suggestions for the ternary operator (`?:`)
 
 ## 1.0.0-beta.12 (July 2, 2024)