+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## next
 
+- Improved suggestions for the ternary operator (`?:`)
+- Disallowed duplicate parameter names in function definitions, i.e., `($a, $a) => expr` is now prohibited
 - Fixed a compilation error for the generated code of `function` expressions in stat mode in some cases
 - Fixed a runtime error for a `comparator function` in some scenarios
-- Improved suggestions for the ternary operator (`?:`)
 
 ## 1.0.0-beta.12 (July 2, 2024)